Fix bug

+++ install.sh
@@ -9,16 +9,18 @@ make all
 make clean
 
 # Install use of library in cygwin environment
-usrlib = /usr/lib/
-usrinc = /usr/include/
+usrlib=/usr/lib/
+usrinc=/usr/include/
 
-echo ""
-echo "Copying lib/libdatabase.a to C++ lib path"
+echo "\nCopying lib/libdatabase.a to C++ lib path"
 cp -f lib/libdatabase.a $usrlib
 
-echo ""
-echo "Copying header files to C++ include path"
+if [! -f lib/libdatabase.a]; then
+    echo "\nInstallation failed.\n"
+    exit 1
+fi
+
+echo "\nCopying header files to C++ include path"
 cp -f include/*.h $usrinc
 
-echo ""
-echo "Install success."
+echo "\nInstallation success."
